+++ lib/rdf/model/triple.ex
@@ -6,7 +6,7 @@ defmodule RDF.Triple do
   RDF values for subject, predicate and object.
   """
 
-  alias RDF.{Statement, PropertyMap}
+  alias RDF.{Statement, BlankNode, PropertyMap}
 
   @type t :: {Statement.subject(), Statement.predicate(), Statement.object()}
 
@@ -92,6 +92,35 @@ defmodule RDF.Triple do
 
   def new({subject, predicate, object, _}, property_map),
     do: new(subject, predicate, object, property_map)
+
+  @doc """
+  Returns a list of all `RDF.BlankNode`s within the given `triple`.
+  """
+  @spec bnodes(t) :: list(BlankNode.t())
+  def bnodes(triple)
+  def bnodes({%BlankNode{} = b, _, %BlankNode{} = b}), do: [b]
+  def bnodes({%BlankNode{} = s, _, %BlankNode{} = o}), do: [s, o]
+  def bnodes({%BlankNode{} = s, _, _}), do: [s]
+  def bnodes({_, _, %BlankNode{} = o}), do: [o]
+  def bnodes(_), do: []
+
+  @doc """
+  Returns whether the given `triple` contains a blank node.
+  """
+  @spec has_bnode?(t) :: boolean
+  def has_bnode?({%BlankNode{}, _, _}), do: true
+  def has_bnode?({_, %BlankNode{}, _}), do: true
+  def has_bnode?({_, _, %BlankNode{}}), do: true
+  def has_bnode?({_, _, _}), do: false
+
+  @doc """
+  Returns whether the given `value` is a component of the given `triple`.
+  """
+  @spec include_value?(t, any) :: boolean
+  def include_value?({value, _, _}, value), do: true
+  def include_value?({_, value, _}, value), do: true
+  def include_value?({_, _, value}, value), do: true
+  def include_value?({_, _, _}), do: false
 
   @doc """
   Returns a tuple of native Elixir values from a `RDF.Triple` of RDF terms.